examples/pytorch-mnist/requirements.txt
pytorch_yard